[SOIN] Extrais les données concernant les niveaux de maturité
