+++ lib-svelte/src/niveaux-maturite/NiveauxMaturite.svelte
@@ -1,10 +1,11 @@
 <script lang="ts">
-  import {niveauxMaturite} from "../test-maturite/TestMaturite.donnees";
+  import { niveauxMaturite } from "./NiveauxMaturite.donnees";
 </script>
+
 <ul>
   {#each niveauxMaturite as niveau}
-  <li>
-    {niveau.label}
-  </li>
-    {/each}
+    <li>
+      {niveau.label}
+    </li>
+  {/each}
 </ul>
